requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.4
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cloudscraper==1.2.71
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
exceptiongroup==1.1.1
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
isort==5.12.0
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
loguru==0.7.0
lxml==4.9.2
more-itertools==8.10.0
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.0
outcome==1.2.0
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
platformdirs==3.5.1
psycopg2-binary==2.9.6
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==5.4.1
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
ssh-import-id==5.11
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
undetected-chromedriver==3.4.7
urllib3==2.0.2
wadllib==1.3.6
webencodings==0.5.1
websockets==11.0.3
wsproto==1.2.0
zenrows==1.3.1
zipp==1.0.0
